OBJParser now has its own interleave thingy
