+++ SluaghEngine/External Dependencies/includes/OBJParser/Parsers.h
@@ -3,7 +3,55 @@
 
 namespace Arf
 {
+	namespace Mesh
+	{
+		namespace InterleaveOption
+		{
+			enum : uint8_t
+			{
+				Position = 1 << 0,
+				TexCoord = 1 << 1,
+				Normal = 1 << 2
+			};
+		}
+		
+		struct Position
+		{
+			float x, y, z;
+		};
 
+		struct TexCoord
+		{
+			float u, v;
+		};
+
+		struct Normal
+		{
+			float x, y, z;
+		};
+
+		struct Vertex
+		{
+			Position pos;
+			Normal norm;
+			TexCoord tex;
+		};
+		struct VertexPositionTexcoord
+		{
+			Position pos;
+			TexCoord tex;
+		};
+		struct VertexPositionNormal
+		{
+			Position pos;
+			Normal norm;
+		};
+		struct Data
+		{
+			uint32_t NumVertices;
+			void* vertices;
+		};
+	}
 
 
 #define PARSER_FILE_NOT_FOUND -1
@@ -15,6 +63,6 @@ namespace Arf
 
 	int ParseArf(const char* filename, ArfData::Data* data, ArfData::DataPointers* dataPointers, uint32_t flags = 0);
 
-	int Interleave(ArfData::Data* data, ArfData::DataPointers* dataPointers, uint32_t flags = 0);
+	int Interleave(ArfData::Data& data, ArfData::DataPointers& dataPointers, Mesh::Data** dataOut, size_t* iSize, uint8_t interleaveOptions = ~0u);
 
 }
